+++ src/c1_foeppl_von_karman/foeppl_von_karman_curvable_bell_elements.h
@@ -916,7 +916,7 @@ namespace oomph
       // this->set_integration_scheme(new_integral_pt);
 
       // Rotated dof helper
-      Rotated_boundary_helper_pt = new RotatedBoundaryHelper(this);
+      Rotated_boundary_helper_pt = new RotatedBoundaryHelper;
     }
 
     /// Destructor: clean up alloacations
@@ -1046,8 +1046,8 @@ namespace oomph
    /// displacement and its normal derivative. We also apply implied
    /// boundary conditions (e.g. specification of dw/dn also implies
    /// d^2w/dn/dzeta etc. boundary_values_pt[i] describes boundary conditions
-   /// for the three displacement components... hierher Aidan: this is where
-   /// it gets complicated: u_x or u_n etc?
+   /// for the three displacement components (i=0,1 in plane (x,y);
+   /// i=2: out-of-plane (z)).
    /// curviline_pt provides a pointer to the representation of the curvilinear
    /// boundary in the triangle mesh.
    virtual void fully_clamp_specified_boundary(
@@ -1059,8 +1059,8 @@ namespace oomph
    /// Pin i.e. pin the in-plane and out of plane displacements only.
    /// We also apply implied boundary conditions (e.g. specification of w
    /// also implies dw/dt etc. boundary_values_pt[i] describes boundary conditions
-   /// for the three displacement components... hierher Aidan: this is where
-   /// it gets complicated: u_x or u_n etc?
+   /// for the three displacement components (i=0,1 in plane (x,y);
+   /// i=2: out-of-plane (z)).
    /// curviline_pt provides a pointer to the representation of the curvilinear
    /// boundary in the triangle mesh.
    virtual void pin_specified_boundary(
@@ -2030,8 +2030,8 @@ namespace oomph
  /// displacement and its normal derivative. We also apply implied
  /// boundary conditions (e.g. specification of dw/dn also implies
  /// d^2w/dn/dzeta etc. boundary_values_pt[i] describes boundary conditions
- /// for the three displacement components... hierher Aidan: this is where
- /// it gets complicated: u_x or u_n etc?
+ /// for the three displacement components (i=0,1 in plane (x,y);
+ /// i=2: out-of-plane (z)).
  /// curviline_pt provides a pointer to the representation of the curvilinear
  /// boundary in the triangle mesh.
  //==========================================================================
@@ -2059,8 +2059,8 @@ namespace oomph
  /// Pin i.e. pin the in-plane and out of plane displacements only.
  /// We also apply implied boundary conditions (e.g. specification of w
  /// also implies dw/dt etc. boundary_values_pt[i] describes boundary conditions
- /// for the three displacement components... hierher Aidan: this is where
- /// it gets complicated: u_x or u_n etc?
+ /// for the three displacement components (i=0,1 in plane (x,y);
+ /// i=2: out-of-plane (z)).
  /// curviline_pt provides a pointer to the representation of the curvilinear
  /// boundary in the triangle mesh.
 //=============================================================================
